+++ projects/_03_independent_cqc/_02_employment_status/fargate/_03_impute.py
@@ -1,4 +1,9 @@
+import projects._03_independent_cqc._02_employment_status.fargate.utils.impute_utils as iUtils
 from polars_utils import utils
+
+EXTRAPOLATION_PERIOD = "2y"
+INTERPOLATION_CAP_PERIOD = "5y"
+ROLLING_AVERAGE_PERIOD = "6mo"
 
 
 def main(
@@ -8,11 +13,26 @@ def main(
     """
     Imputes values to fill gaps between and around known employment status values.
 
+    Short-term gaps in each location and job role's percentages are filled first, then
+    those percentages are averaged over a rolling window per primary service type, region
+    and job role.
+
     Args:
         cleaned_data_source (str): path to the cleaned data
         imputed_data_destination (str): destination for output
     """
     lf = utils.scan_parquet(cleaned_data_source)
+
+    lf = iUtils.add_short_term_imputed_percentages(
+        lf,
+        extrapolation_period=EXTRAPOLATION_PERIOD,
+        interpolation_cap_period=INTERPOLATION_CAP_PERIOD,
+    )
+
+    lf = iUtils.add_rolling_average_percentages(
+        lf,
+        rolling_period=ROLLING_AVERAGE_PERIOD,
+    )
 
     utils.sink_to_parquet(
         lazy_df=lf,